requirements.txt
fastapi~=0.115.0
uvicorn==0.31.0
mysql-connector-python==9.0.0